refactor so card component takes arbitrary children
